+++ NewLife.Core/Data/DbRow.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace NewLife.Data
+{
+    /// <summary>数据行</summary>
+    public class DbRow
+    {
+    }
+}
