+++ XCode/Entity/IEntitySession.cs
@@ -73,7 +73,7 @@ namespace XCode
         /// <param name="startRowIndex">开始行，0表示第一行</param>
         /// <param name="maximumRows">最大返回行数，0表示所有行</param>
         /// <returns></returns>
-        DbSet Query(SelectBuilder builder, Int64 startRowIndex, Int64 maximumRows);
+        DbTable Query(SelectBuilder builder, Int64 startRowIndex, Int64 maximumRows);
 
         /// <summary>查询记录数</summary>
         /// <param name="builder">查询生成器</param>
